refactor: 오목 도메인 로직 모듈 분리
