--- app/src/main/java/woowacourse/omok/src/main/kotlin/omok/Application.kt
@@ -1,10 +1,0 @@
-package woowacourse.omok.src.main.kotlin.omok
-
-import woowacourse.omok.src.main.kotlin.omok.controller.OmokController
-import woowacourse.omok.src.main.kotlin.omok.view.InputView
-import woowacourse.omok.src.main.kotlin.omok.view.OutputView
-
-fun main() {
-    val controller = OmokController(InputView(), OutputView())
-    controller.start()
-}
