--- app/src/main/java/woowacourse/omok/src/main/kotlin/omok/model/stone/BlackStone.kt
@@ -1,5 +1,0 @@
-package woowacourse.omok.src.main.kotlin.omok.model.stone
-
-object BlackStone : GoStone() {
-    override val stoneType = StoneType.BLACK_STONE
-}
