--- app/src/main/java/woowacourse/omok/src/main/kotlin/omok/model/stone/StoneType.kt
@@ -1,7 +1,0 @@
-package woowacourse.omok.src.main.kotlin.omok.model.stone
-
-enum class StoneType {
-    BLACK_STONE,
-    WHITE_STONE,
-    NONE,
-}
