--- app/src/main/java/woowacourse/omok/src/main/kotlin/omok/model/stone/WhiteStone.kt
@@ -1,5 +1,0 @@
-package woowacourse.omok.src.main.kotlin.omok.model.stone
-
-object WhiteStone : GoStone() {
-    override val stoneType = StoneType.WHITE_STONE
-}
