--- app/src/test/java/woowacourse/omok/kotlin/omok/model/stone/BlackStoneTestType.kt
@@ -1,32 +1,0 @@
-package woowacourse.omok.kotlin.omok.model.stone
-
-import org.assertj.core.api.Assertions.assertThatThrownBy
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertDoesNotThrow
-import woowacourse.omok.kotlin.omok.model.resetBoard
-import woowacourse.omok.src.main.kotlin.omok.model.position.Position
-import woowacourse.omok.src.main.kotlin.omok.model.stone.BlackStone
-
-class BlackStoneTestType {
-    @BeforeEach
-    fun setUp() {
-        resetBoard()
-    }
-
-    @Test
-    fun `이미 돌이 놓인 자리에 돌을 놓을 경우 예외가 발생한다`() {
-        val stone = BlackStone
-        val position = Position(0, 0)
-        stone.putStone(position)
-        assertThatThrownBy { stone.putStone(position) }.isExactlyInstanceOf(IllegalArgumentException::class.java)
-            .hasMessage("이미 놓여진 자리입니다.\n")
-    }
-
-    @Test
-    fun `유효한 자리에 돌을 놓을 경우 예외가 발생하지 않는다`() {
-        val stone = BlackStone
-        val position = Position(1, 2)
-        assertDoesNotThrow { stone.putStone(position) }
-    }
-}
