+++ build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
     id("com.android.application") version "8.3.1" apply false
     id("org.jetbrains.kotlin.android") version "1.9.0" apply false
     id("org.jlleitschuh.gradle.ktlint") version "12.1.0" apply false
+    id("org.jetbrains.kotlin.jvm") version "1.9.0" apply false
 }
 
 allprojects {
