+++ model/src/main/kotlin/omok/model/stone/BlackStone.kt
@@ -1,0 +1,5 @@
+package omok.model.stone
+
+object BlackStone : GoStone() {
+    override val stoneType = StoneType.BLACK_STONE
+}
