+++ model/src/main/kotlin/omok/model/stone/StoneType.kt
@@ -1,0 +1,7 @@
+package omok.model.stone
+
+enum class StoneType {
+    BLACK_STONE,
+    WHITE_STONE,
+    NONE,
+}
