+++ model/src/main/kotlin/omok/model/stone/WhiteStone.kt
@@ -1,0 +1,5 @@
+package omok.model.stone
+
+object WhiteStone : GoStone() {
+    override val stoneType = StoneType.WHITE_STONE
+}
