+++ model/src/test/kotlin/omok/model/position/RelativeDirectionTest.kt
@@ -1,0 +1,19 @@
+package omok.model.position
+
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+
+class RelativeDirectionTest {
+    @Test
+    fun `돌의 위치 값이 반전되는지 확인한다`() {
+        // given
+        val position = RelativeDirection(1, 1)
+
+        // when
+        val actual = -position
+        val expected = RelativeDirection(-1, -1)
+
+        // then
+        assertThat(actual).isEqualTo(expected)
+    }
+}
