+++ settings.gradle.kts
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "kotlin-omok"
 include(":app")
+include(":model")
